+++ test/runtests.jl
@@ -4,6 +4,7 @@ using LinearAlgebra
 
 @testset "Test everything" begin
     include("test_aqua.jl")
+    include("test_readme.jl")
     include("test_golden.jl")
     include("test_convergence.jl")
     include("test_invariants.jl")